+++ app/post/PostRepository.scala
@@ -1,4 +1,6 @@
-package com.lightbend.blog.post
+package post
+
+import scala.concurrent.Future
 
 final case class PostData(id: PostId, title: String, body: String)
 
@@ -11,4 +13,15 @@ object PostId {
     require(raw != null)
     new PostId(Integer.parseInt(raw))
   }
+}
+
+/**
+ * A pure non-blocking interface for the PostRepository.
+ */
+trait PostRepository {
+  def create(data: PostData): Future[PostId]
+
+  def list(): Future[Iterable[PostData]]
+
+  def get(id: PostId): Future[Option[PostData]]
 }
