--- app/views/index.scala.html
@@ -1,11 +1,0 @@
-@(router: post.PostRouter)(implicit request: Request[_])
-
-@main("REST API") {
-  <div class="header">
-    <h1 class="header-title">REST API</h1>
-  </div>
-
-  <ul>
-    <li><a href="@{router.prefix}">@{router.prefix}</a></li>
-  </ul>
-}
