--- app/views/main.scala.html
@@ -1,30 +1,0 @@
-@(title: String)(content: Html)(implicit r: Request[_])
-
-<!DOCTYPE html>
-
-<html>
-  <head>
-    <title>@title</title>
-    <link rel="stylesheet" media="screen" href="@routes.Assets.at("lib/bootstrap/css/bootstrap.css")">
-    <link rel="stylesheet" media="screen" href="@routes.Assets.at("stylesheets/main.css")">
-  </head>
-  <body>
-    <div class="navbar navbar-default navbar-fixed-top">
-      <div class="navbar-header">
-        <a class="navbar-brand" href="@routes.HomeController.index">REST API</a>
-      </div>
-    </div>
-
-    <div class="container">
-    @content
-    </div>
-
-    <div class="footer">
-      <div class="container">
-        <p></p>
-      </div>
-    </div>
-
-    <script src="@routes.Assets.at("lib/jquery/jquery.min.js")" type="text/javascript"></script>
-  </body>
-</html>
