--- app/views/posts/flash.scala.html
@@ -1,11 +1,0 @@
-@()(implicit request: post.PostRequest[_])
-
-@request.flashSuccess.map { message =>
-  <div class="alert alert-success" role="alert">@message</div>
-}
-
-@request.flashError.map { message =>
-  <div class="alert alert-danger" role="alert">@message</div>
-}
-
-
