--- app/views/posts/index.scala.html
@@ -1,28 +1,0 @@
-@(posts: Iterable[post.PostResource], form: Form[post.PostFormInput])(implicit request: post.PostRequest[_])
-
-@main("Posts") {
-
-  <div class="header">
-    <h1 class="header-title">Posts</h1>
-  </div>
-
-  @flash()
-
-  <table class="table table-striped">
-    <thead>
-      <td>Link</td>
-      <td>Title</td>
-      <td>Body</td>
-      <td>Comments</td>
-    </thead>
-    <tbody>
-    @for(post <- posts) {
-      @row(post)
-    }
-    </tbody>
-  </table>
-
-  <div>
-    @create(form)
-  </div>
-}
