--- app/views/posts/main.scala.html
@@ -1,3 +1,0 @@
-@(title: String)(content: Html)(implicit request: post.PostRequest[_])
-
-@views.html.main(title)(content)
