--- app/views/posts/row.scala.html
@@ -1,24 +1,0 @@
-@(p: post.PostResource)(implicit request: post.PostRequest[_])
-
-<tr>
-
-  <td>
-    <a href="@{p.link}">@{p.link}</a>
-  </td>
-
-  <td>
-    @{p.title}
-  </td>
-
-  <td>
-    @{p.body}
-  </td>
-
-  <td>
-    @for(c <- p.comments) {
-      @{c.body}
-      <br>
-    }
-  </td>
-
-</tr>
