--- app/views/posts/show.scala.html
@@ -1,35 +1,0 @@
-@(p: post.PostResource)(implicit request: post.PostRequest[_])
-
-@main("Posts") {
-
-  <div class="header">
-    <h1 class="header-title">Post @{p.id}</h1>
-  </div>
-
-  <div class="row">
-    <div class="col-md-2"><strong>ID</strong></div>
-    <div class="col-md-6">@{p.id}</div>
-  </div>
-  <div class="row">
-    <div class="col-md-2"><strong>Link</strong></div>
-    <div class="col-md-6">@{p.link}</div>
-  </div>
-  <div class="row">
-    <div class="col-md-2"><strong>Title</strong></div>
-    <div class="col-md-6">@{p.title}</div>
-  </div>
-  <div class="row">
-    <div class="col-md-2"><strong>Body</strong></div>
-    <div class="col-md-6">@{p.body}</div>
-  </div>
-  <div class="row">
-    <div class="col-md-2"><strong>Comment</strong></div>
-    <div class="col-md-6">
-      @for(c <- p.comments) {
-        @{c.body}
-        <br>
-      }
-    </div>
-  </div>
-
-}
