--- disabledAlgorithms.properties
@@ -1,4 +1,0 @@
-# disabledAlgorithms.properties
-# http://sim.ivi.co/2013/11/harness-ssl-and-jsse-key-size-control.html
-jdk.tls.disabledAlgorithms=SSLv3, TLSv1, TLSv1.1, EC keySize < 160, RSA keySize < 2048, DSA keySize < 2048
-jdk.certpath.disabledAlgorithms=MD2, MD4, MD5, EC keySize < 160, RSA keySize < 2048, DSA keySize < 2048
