--- test/ApplicationSpec.scala
@@ -1,23 +1,0 @@
-import org.scalatestplus.play._
-import play.api.test._
-import play.api.test.Helpers._
-
-class ApplicationSpec extends PlaySpec with OneAppPerTest {
-
-  "Routes" should {
-    "send 404 on a bad request" in  {
-      route(app, FakeRequest(GET, "/boum")).map(status(_)) mustBe Some(NOT_FOUND)
-    }
-  }
-
-  "HomeController" should {
-    "render the index page" in {
-      val home = route(app, FakeRequest(GET, "/")).get
-
-      status(home) mustBe OK
-      contentType(home) mustBe Some("text/html")
-      contentAsString(home) must include ("Your new application is ready.")
-    }
-  }
-
-}
